Primer commit: subir proyecto inicial

+++ backend/src/main/java/FitStore/TpoGrupo10/service/impl/UsuarioServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 
 @Service
 public class UsuarioServiceImpl implements UsuarioService {
@@ -24,12 +25,14 @@ public class UsuarioServiceImpl implements UsuarioService {
 
     @Override
     public UsuarioModel findById(Long id) {
-        return repository.findById(id).orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+        return repository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Usuario no encontrado con ID: " + id));
     }
 
     @Override
     public UsuarioModel findByEmail(String email) {
-        return repository.findByEmail(email).orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+        return repository.findByEmail(email)
+                .orElseThrow(() -> new EntityNotFoundException("Usuario no encontrado con email: " + email));
     }
 
     @Override
@@ -39,8 +42,9 @@ public class UsuarioServiceImpl implements UsuarioService {
 
     @Override
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new EntityNotFoundException("Usuario no encontrado con ID: " + id);
+        }
         repository.deleteById(id);
     }
-
-
 }
